Extract SyncCoordinator from AppState
Move all sync orchestration into a dedicated SyncCoordinator class:
- roadflareSyncStore ownership and markDirty forwarding
- All 4 domain resolution methods (profile, drivers, profileBackup, rideHistory)
- Publish methods (publishProfile, publishProfileBackup, saveAndPublishSettings)
- buildProfileBackupContent and preserveProfileBackupSettingsTemplate
- performStartupSync with typed SyncProgress callback for import UI
- flushPendingSyncPublishes(rideCoordinator:) for reconnect
- wireTrackingCallbacks consolidates all 5 callback wirings (settings + repos)
- teardown detaches ALL callbacks before any clearAll, preventing the
  stale-callback sync regression class of bug

AppState becomes a thin shell: 727 -> 392 lines (46% reduction).
Zero view file changes. Part 2/3 of #18.
