+++ RoadFlare/RoadFlare/ViewModels/AppState.swift
@@ -13,6 +13,9 @@ enum AuthState {
 }
 
 /// Central app state coordinator. Owns SDK services and manages auth lifecycle.
+///
+/// Views access this via `@Environment(AppState.self)`. All sync orchestration
+/// is delegated to `SyncCoordinator`; connection watchdog to `ConnectionCoordinator`.
 @Observable
 @MainActor
 final class AppState {
@@ -54,26 +57,14 @@ final class AppState {
     var syncRestoredLocations: Int = 0
     var syncRestoredName: Bool = false
 
-    // MARK: - Storage
+    // MARK: - Private Coordinators & Storage
 
     private let keychainStorage = KeychainStorage(service: "com.roadflare.keys")
     private let driversPersistence = UserDefaultsDriversPersistence()
-    private var roadflareSyncStore: RoadflareSyncStateStore?
+    private var syncCoordinator: SyncCoordinator?
     private let connectionCoordinator = ConnectionCoordinator()
-    private var isPublishingProfileBackup = false
-    private var profileBackupRepublishRequested = false
-    private var profileBackupSettingsTemplate = SettingsBackupContent()
 
     // MARK: - Init
-
-    init() {
-        settings.onProfileChanged = { [weak self] in
-            self?.roadflareSyncStore?.markDirty(.profile)
-        }
-        settings.onProfileBackupChanged = { [weak self] in
-            self?.roadflareSyncStore?.markDirty(.profileBackup)
-        }
-    }
 
     private static let hasLaunchedKey = "roadflare_has_launched"
 
@@ -134,105 +125,49 @@ final class AppState {
 
         // Navigate based on what was restored
         if settings.profileName.isEmpty {
-            // No name → go to profile setup
             authState = .profileIncomplete
         } else if settings.roadflarePaymentMethods.isEmpty {
-            // Have name but no payment methods → go to payment setup
             authState = .paymentSetup
         } else {
-            // Have both → fully ready
             settings.profileCompleted = true
             authState = .ready
         }
     }
 
     /// Mark profile name as set, publish Kind 0 to Nostr, move to payment setup.
-    /// Does NOT publish Kind 30177 yet — payment methods aren't set until next step.
     func completeProfileSetup(name: String) async {
         settings.profileName = name
-        roadflareSyncStore?.markDirty(.profile)
-        await publishProfile()
+        syncCoordinator?.markDirty(.profile)
+        await syncCoordinator?.publishProfile()
         authState = .paymentSetup
     }
 
     /// Mark payment setup as done, finish onboarding. Publishes profile + settings backup.
     func completePaymentSetup() async {
         settings.profileCompleted = true
-        roadflareSyncStore?.markDirty(.profileBackup)
-        await saveAndPublishSettings()
+        syncCoordinator?.markDirty(.profileBackup)
+        await syncCoordinator?.saveAndPublishSettings()
         authState = .ready
     }
 
-    /// Publish Kind 0 metadata to Nostr.
-    func publishProfile() async {
-        guard let service = roadflareDomainService else { return }
-        let profile = UserProfileContent(
-            name: settings.profileName,
-            displayName: settings.profileName
-        )
-        do {
-            let event = try await service.publishProfile(profile)
-            roadflareSyncStore?.markPublished(.profile, at: event.createdAt)
-            AppLogger.auth.info("Published profile to Nostr")
-        } catch {
-            AppLogger.auth.info("Failed to publish profile: \(error)")
-        }
-    }
+    // MARK: - Forwarding to SyncCoordinator
 
-    /// Publish Kind 30177 encrypted profile backup (settings + saved locations) to Nostr.
-    func publishProfileBackup() async {
-        guard let service = roadflareDomainService else { return }
-        if isPublishingProfileBackup {
-            profileBackupRepublishRequested = true
-            return
-        }
-
-        isPublishingProfileBackup = true
-        defer { isPublishingProfileBackup = false }
-
-        repeat {
-            profileBackupRepublishRequested = false
-            let backup = buildProfileBackupContent()
-            do {
-                let event = try await service.publishProfileBackup(backup)
-                roadflareSyncStore?.markPublished(.profileBackup, at: event.createdAt)
-                AppLogger.auth.info("Published profile backup to Nostr")
-            } catch {
-                AppLogger.auth.info("Failed to publish profile backup: \(error)")
-            }
-        } while profileBackupRepublishRequested
-    }
+    func publishProfile() async { await syncCoordinator?.publishProfile() }
+    func publishProfileBackup() async { await syncCoordinator?.publishProfileBackup() }
+    func saveAndPublishSettings() async { await syncCoordinator?.saveAndPublishSettings() }
 
     func buildProfileBackupContent() -> ProfileBackupContent {
-        var settingsBackup = profileBackupSettingsTemplate
-        settingsBackup.roadflarePaymentMethods = settings.roadflarePaymentMethods
-
-        return ProfileBackupContent(
-            savedLocations: savedLocations.locations.map { loc in
-                SavedLocationBackup(
-                    displayName: loc.displayName, lat: loc.latitude, lon: loc.longitude,
-                    addressLine: loc.addressLine,
-                    isPinned: loc.isPinned,
-                    nickname: loc.nickname,
-                    timestampMs: loc.timestampMs
-                )
-            },
-            settings: settingsBackup
-        )
+        syncCoordinator?.buildProfileBackupContent()
+            ?? ProfileBackupContent(savedLocations: [], settings: SettingsBackupContent())
     }
 
-    func preserveProfileBackupSettingsTemplate(_ settings: SettingsBackupContent) {
-        profileBackupSettingsTemplate = settings
+    func preserveProfileBackupSettingsTemplate(_ s: SettingsBackupContent) {
+        syncCoordinator?.preserveProfileBackupSettingsTemplate(s)
     }
 
-    /// Publish both Kind 0 (public profile) and Kind 30177 (encrypted backup) to Nostr.
-    func saveAndPublishSettings() async {
-        await publishProfile()
-        await publishProfileBackup()
-    }
+    // MARK: - Driver Key Management
 
     /// Try to restore a specific driver's key from our Kind 30011 backup on the relay.
-    /// Used during mid-session re-add when the key was lost locally but exists in the backup.
     func restoreKeyFromBackup(driverPubkey: String) async {
         guard let service = roadflareDomainService,
               let repo = driversRepository else { return }
@@ -260,7 +195,6 @@ final class AppState {
     }
 
     /// Send Kind 3187 follow notification to a driver (real-time nudge).
-    /// The rider's Kind 30011 p-tags are the source of truth — this is just a push notification.
     func sendFollowNotification(driverPubkey: String) async {
         guard let kp = keypair, let rm = relayManager,
               !settings.profileName.isEmpty else { return }
@@ -277,11 +211,23 @@ final class AppState {
         }
     }
 
+    // MARK: - Connection & Foreground
+
     /// Called when app returns to foreground. Reconnects relays and restarts subscriptions if needed.
     func handleForeground() async {
         guard authState == .ready else { return }
         await reconnectAndRestoreSession()
     }
+
+    func reconnectAndRestoreSession() async {
+        guard let rm = relayManager else { return }
+        await rm.reconnectIfNeeded()
+        guard await rm.isConnected else { return }
+        await syncCoordinator?.flushPendingSyncPublishes(rideCoordinator: rideCoordinator)
+        await rideCoordinator?.restoreLiveSubscriptions()
+    }
+
+    // MARK: - Auth Helpers
 
     func resolveLocalAuthState() -> AuthState {
         if settings.profileName.isEmpty {
@@ -301,243 +247,24 @@ final class AppState {
         authState = .loggedOut
     }
 
-    private func detachRepositoryTrackingCallbacks() {
-        driversRepository?.onDriversChanged = nil
-        rideHistory.onRidesChanged = nil
-        savedLocations.onChange = nil
-        savedLocations.onFavoritesChanged = nil
-    }
-
-    // MARK: - Private
-
-    /// Fetch, resolve, and apply RoadFlare startup state using SDK-owned helpers.
-    private func performStartupSync(repo: FollowedDriversRepository, importFlow: Bool) async {
-        guard let service = roadflareDomainService else { return }
-
-        let remote = await service.fetchStartupRemoteState()
-
-        if importFlow { syncStatus = "Restoring your profile..." }
-        await applyProfileResolution(remote: remote.profile)
-        if importFlow { syncRestoredName = !settings.profileName.isEmpty }
-
-        if importFlow { syncStatus = "Restoring your drivers..." }
-        await applyFollowedDriversResolution(repo: repo, remote: remote.followedDrivers)
-        if importFlow { syncRestoredDrivers = repo.drivers.count }
-
-        if importFlow { syncStatus = "Restoring settings..." }
-        await applyProfileBackupResolution(remote: remote.profileBackup)
-        if importFlow { syncRestoredLocations = savedLocations.locations.count }
-
-        if importFlow { syncStatus = "Restoring ride history..." }
-        await applyRideHistoryResolution(remote: remote.rideHistory)
-
-        if importFlow { syncStatus = "Loading driver info..." }
-        let driverProfiles = await service.fetchDriverProfiles(pubkeys: repo.allPubkeys)
-        for (pubkey, snapshot) in driverProfiles {
-            repo.cacheDriverProfile(pubkey: pubkey, profile: snapshot.value)
-        }
-        if !driverProfiles.isEmpty {
-            AppLogger.auth.info("Fetched profiles for \(driverProfiles.count) driver(s)")
-        }
-
-        if importFlow {
-            syncStatus = "Done!"
-            try? await Task.sleep(for: .seconds(1))
-        }
-    }
-
-    private func applyProfileResolution(
-        remote: RoadflareDomainService.StartupRemoteDomain<UserProfileContent>
-    ) async {
-        guard let syncStore = roadflareSyncStore else { return }
-        let metadata = syncStore.metadata(for: .profile)
-        let resolution = RoadflareDomainService.resolve(
-            domain: .profile,
-            metadata: metadata,
-            remoteCreatedAt: remote.latestSeenCreatedAt
-        )
-        let shouldSeedLegacyLocal = RoadflareDomainService.shouldSeedLegacyLocalState(
-            metadata: metadata,
-            remoteCreatedAt: remote.latestSeenCreatedAt,
-            hasLocalState: !settings.profileName.isEmpty
-        )
-
-        if resolution.source == .remote,
-           let snapshot = remote.snapshot,
-           snapshot.createdAt == remote.latestSeenCreatedAt {
-            let name = snapshot.value.displayName ?? snapshot.value.name ?? ""
-            settings.performWithoutChangeTracking {
-                settings.profileName = name
-            }
-            if !name.isEmpty {
-                AppLogger.auth.info("Restored profile name from Nostr: \(name)")
-            }
-            syncStore.markPublished(.profile, at: snapshot.createdAt)
-        } else if resolution.source == .remote, remote.latestSeenCreatedAt != nil {
-            AppLogger.auth.warning("Latest profile metadata is not decodable; preserving local profile state")
-        } else if resolution.shouldPublishLocal || shouldSeedLegacyLocal {
-            if shouldSeedLegacyLocal {
-                syncStore.markDirty(.profile)
-            }
-            await publishProfile()
-        }
-    }
-
-    private func applyFollowedDriversResolution(
-        repo: FollowedDriversRepository,
-        remote: RoadflareDomainService.StartupRemoteDomain<FollowedDriversContent>
-    ) async {
-        guard let service = roadflareDomainService else { return }
-        guard let syncStore = roadflareSyncStore else { return }
-        let metadata = syncStore.metadata(for: .followedDrivers)
-        let resolution = RoadflareDomainService.resolve(
-            domain: .followedDrivers,
-            metadata: metadata,
-            remoteCreatedAt: remote.latestSeenCreatedAt
-        )
-        let shouldSeedLegacyLocal = RoadflareDomainService.shouldSeedLegacyLocalState(
-            metadata: metadata,
-            remoteCreatedAt: remote.latestSeenCreatedAt,
-            hasLocalState: repo.hasDrivers
-        )
-
-        if resolution.source == .remote,
-           let snapshot = remote.snapshot,
-           snapshot.createdAt == remote.latestSeenCreatedAt {
-            repo.restoreFromNostr(content: snapshot.value)
-            syncStore.markPublished(.followedDrivers, at: snapshot.createdAt)
-            if !snapshot.value.drivers.isEmpty {
-                AppLogger.auth.info("Restored \(snapshot.value.drivers.count) drivers from Nostr")
-            }
-        } else if resolution.source == .remote, remote.latestSeenCreatedAt != nil {
-            AppLogger.auth.warning("Latest followed-drivers snapshot is not decodable; preserving local driver state")
-        } else if resolution.shouldPublishLocal || shouldSeedLegacyLocal {
-            if shouldSeedLegacyLocal {
-                syncStore.markDirty(.followedDrivers)
-            }
-            do {
-                let event = try await service.publishFollowedDriversList(repo.drivers)
-                syncStore.markPublished(.followedDrivers, at: event.createdAt)
-                AppLogger.auth.info("Published followed drivers list to Nostr")
-            } catch {
-                AppLogger.auth.info("Failed to publish followed drivers list: \(error)")
-            }
-        }
-    }
-
-    private func applyProfileBackupResolution(
-        remote: RoadflareDomainService.StartupRemoteDomain<ProfileBackupContent>
-    ) async {
-        guard roadflareDomainService != nil else { return }
-        guard let syncStore = roadflareSyncStore else { return }
-        if let snapshot = remote.snapshot {
-            preserveProfileBackupSettingsTemplate(snapshot.value.settings)
-        }
-        let metadata = syncStore.metadata(for: .profileBackup)
-        let resolution = RoadflareDomainService.resolve(
-            domain: .profileBackup,
-            metadata: metadata,
-            remoteCreatedAt: remote.latestSeenCreatedAt
-        )
-        let shouldSeedLegacyLocal = RoadflareDomainService.shouldSeedLegacyLocalState(
-            metadata: metadata,
-            remoteCreatedAt: remote.latestSeenCreatedAt,
-            hasLocalState: !settings.roadflarePaymentMethods.isEmpty
-                || !savedLocations.locations.isEmpty
-        )
-
-        if resolution.source == .remote,
-           let snapshot = remote.snapshot,
-           snapshot.createdAt == remote.latestSeenCreatedAt {
-            applyRemoteProfileBackup(snapshot.value)
-            syncStore.markPublished(.profileBackup, at: snapshot.createdAt)
-        } else if resolution.source == .remote, remote.latestSeenCreatedAt != nil {
-            AppLogger.auth.warning("Latest profile backup is not decodable; preserving local backup state")
-        } else if resolution.shouldPublishLocal || shouldSeedLegacyLocal {
-            if shouldSeedLegacyLocal {
-                syncStore.markDirty(.profileBackup)
-            }
-            await publishProfileBackup()
-        }
-    }
-
-    private func applyRemoteProfileBackup(_ backup: ProfileBackupContent) {
-        preserveProfileBackupSettingsTemplate(backup.settings)
-        settings.performWithoutChangeTracking {
-            settings.setRoadflarePaymentMethods(backup.settings.roadflarePaymentMethods)
-        }
-
-        savedLocations.restoreFromBackup(backup.savedLocations.map { loc in
-            SavedLocation(
-                latitude: loc.lat,
-                longitude: loc.lon,
-                displayName: loc.displayName,
-                addressLine: loc.addressLine ?? loc.displayName,
-                isPinned: loc.isPinned,
-                nickname: loc.nickname,
-                timestampMs: loc.timestampMs ?? Int(Date.now.timeIntervalSince1970 * 1000)
-            )
-        })
-
-        if !backup.savedLocations.isEmpty {
-            AppLogger.auth.info("Restored \(backup.savedLocations.count) saved locations")
-        }
-    }
-
-    private func applyRideHistoryResolution(
-        remote: RoadflareDomainService.StartupRemoteDomain<RideHistoryBackupContent>
-    ) async {
-        guard let service = roadflareDomainService else { return }
-        guard let syncStore = roadflareSyncStore else { return }
-        let metadata = syncStore.metadata(for: .rideHistory)
-        let resolution = RoadflareDomainService.resolve(
-            domain: .rideHistory,
-            metadata: metadata,
-            remoteCreatedAt: remote.latestSeenCreatedAt
-        )
-        let shouldSeedLegacyLocal = RoadflareDomainService.shouldSeedLegacyLocalState(
-            metadata: metadata,
-            remoteCreatedAt: remote.latestSeenCreatedAt,
-            hasLocalState: !rideHistory.rides.isEmpty
-        )
-
-        if resolution.source == .remote,
-           let snapshot = remote.snapshot,
-           snapshot.createdAt == remote.latestSeenCreatedAt {
-            // Remote is authoritative — full replace (matches other domain patterns)
-            rideHistory.restoreFromBackup(snapshot.value.rides)
-            syncStore.markPublished(.rideHistory, at: snapshot.createdAt)
-            if !snapshot.value.rides.isEmpty {
-                AppLogger.auth.info("Restored \(snapshot.value.rides.count) ride(s) from Nostr")
-            }
-        } else if (resolution.shouldPublishLocal || shouldSeedLegacyLocal), !rideHistory.rides.isEmpty {
-            if shouldSeedLegacyLocal {
-                syncStore.markDirty(.rideHistory)
-            }
-            do {
-                let content = RideHistoryBackupContent(rides: rideHistory.rides)
-                let event = try await service.publishRideHistoryBackup(content)
-                syncStore.markPublished(.rideHistory, at: event.createdAt)
-                AppLogger.auth.info("Published ride history backup to Nostr")
-            } catch {
-                AppLogger.auth.info("Failed to publish ride history backup: \(error)")
-            }
-        }
-    }
+    // MARK: - Service Setup
 
     /// Setup with sync status updates for the import flow UI.
     private func setupServicesWithSync(keypair: NostrKeypair) async {
         let rm = RelayManager(keypair: keypair)
         self.relayManager = rm
+
+        let sync = SyncCoordinator(settings: settings, savedLocations: savedLocations, rideHistory: rideHistory)
         let syncStore = RoadflareSyncStateStore(namespace: keypair.publicKeyHex)
-        self.roadflareSyncStore = syncStore
+        let service = RoadflareDomainService(relayManager: rm, keypair: keypair)
+        sync.configure(syncStore: syncStore, domainService: service)
+        self.syncCoordinator = sync
+        self.roadflareDomainService = service
+
         let repo = FollowedDriversRepository(persistence: driversPersistence)
         self.driversRepository = repo
-        configureDriversRepositoryTracking(repo, syncStore: syncStore)
-        configureRideHistoryTracking(syncStore: syncStore)
-        configureSavedLocationsTracking(syncStore: syncStore)
-        let service = RoadflareDomainService(relayManager: rm, keypair: keypair)
-        self.roadflareDomainService = service
+        sync.wireTrackingCallbacks(driversRepo: repo)
+
         self.fareCalculator = FareCalculator()
         self.remoteConfigManager = RemoteConfigManager(relayManager: rm)
         bitcoinPrice.start()
@@ -550,14 +277,21 @@ final class AppState {
             try? await Task.sleep(for: .seconds(1))
         }
 
-        await performStartupSync(repo: repo, importFlow: true)
+        await sync.performStartupSync(repo: repo, importFlow: true) { [weak self] progress in
+            switch progress {
+            case .status(let msg): self?.syncStatus = msg
+            case .profileRestored(let found): self?.syncRestoredName = found
+            case .driversRestored(let count): self?.syncRestoredDrivers = count
+            case .locationsRestored(let count): self?.syncRestoredLocations = count
+            }
+        }
 
         let coordinator = RideCoordinator(
             relayManager: rm, keypair: keypair,
             driversRepository: repo, settings: settings,
             rideHistory: rideHistory, bitcoinPrice: bitcoinPrice,
             roadflareDomainService: service,
-            roadflareSyncStore: syncStore
+            roadflareSyncStore: sync.roadflareSyncStore
         )
         self.rideCoordinator = coordinator
         await coordinator.restoreLiveSubscriptions()
@@ -572,15 +306,18 @@ final class AppState {
     private func setupServices(keypair: NostrKeypair) async {
         let rm = RelayManager(keypair: keypair)
         self.relayManager = rm
+
+        let sync = SyncCoordinator(settings: settings, savedLocations: savedLocations, rideHistory: rideHistory)
         let syncStore = RoadflareSyncStateStore(namespace: keypair.publicKeyHex)
-        self.roadflareSyncStore = syncStore
+        let service = RoadflareDomainService(relayManager: rm, keypair: keypair)
+        sync.configure(syncStore: syncStore, domainService: service)
+        self.syncCoordinator = sync
+        self.roadflareDomainService = service
+
         let repo = FollowedDriversRepository(persistence: driversPersistence)
         self.driversRepository = repo
-        configureDriversRepositoryTracking(repo, syncStore: syncStore)
-        configureRideHistoryTracking(syncStore: syncStore)
-        configureSavedLocationsTracking(syncStore: syncStore)
-        let service = RoadflareDomainService(relayManager: rm, keypair: keypair)
-        self.roadflareDomainService = service
+        sync.wireTrackingCallbacks(driversRepo: repo)
+
         self.fareCalculator = FareCalculator()
         self.remoteConfigManager = RemoteConfigManager(relayManager: rm)
         bitcoinPrice.start()
@@ -594,14 +331,14 @@ final class AppState {
             AppLogger.auth.info(" Relay connection FAILED: \(error)")
         }
 
-        await performStartupSync(repo: repo, importFlow: false)
+        await sync.performStartupSync(repo: repo, importFlow: false)
 
         let coordinator = RideCoordinator(
             relayManager: rm, keypair: keypair,
             driversRepository: repo, settings: settings,
             rideHistory: rideHistory, bitcoinPrice: bitcoinPrice,
             roadflareDomainService: service,
-            roadflareSyncStore: syncStore
+            roadflareSyncStore: sync.roadflareSyncStore
         )
         self.rideCoordinator = coordinator
         AppLogger.auth.info("Starting subscriptions... (\(repo.drivers.count) drivers loaded)")
@@ -614,75 +351,21 @@ final class AppState {
         )
     }
 
-    private func configureDriversRepositoryTracking(
-        _ repo: FollowedDriversRepository,
-        syncStore: RoadflareSyncStateStore?
-    ) {
-        repo.onDriversChanged = { source in
-            guard source == .local else { return }
-            syncStore?.markDirty(.followedDrivers)
-        }
-    }
-
-    private func configureRideHistoryTracking(syncStore: RoadflareSyncStateStore?) {
-        rideHistory.onRidesChanged = {
-            syncStore?.markDirty(.rideHistory)
-        }
-    }
-
-    private func configureSavedLocationsTracking(syncStore: RoadflareSyncStateStore?) {
-        savedLocations.onChange = {
-            syncStore?.markDirty(.profileBackup)
-        }
-    }
-
-    func reconnectAndRestoreSession() async {
-        guard let rm = relayManager else { return }
-        await rm.reconnectIfNeeded()
-        guard await rm.isConnected else { return }
-        await flushPendingSyncPublishes()
-        await rideCoordinator?.restoreLiveSubscriptions()
-    }
-
-    private func flushPendingSyncPublishes() async {
-        guard let rm = relayManager,
-              let syncStore = roadflareSyncStore,
-              await rm.isConnected else { return }
-
-        if syncStore.metadata(for: .profile).isDirty {
-            await publishProfile()
-        }
-        if syncStore.metadata(for: .followedDrivers).isDirty {
-            await rideCoordinator?.publishFollowedDriversList()
-        }
-        if syncStore.metadata(for: .profileBackup).isDirty {
-            await publishProfileBackup()
-        }
-        if syncStore.metadata(for: .rideHistory).isDirty, !rideHistory.rides.isEmpty {
-            do {
-                let content = RideHistoryBackupContent(rides: rideHistory.rides)
-                if let service = roadflareDomainService {
-                    let event = try await service.publishRideHistoryBackup(content)
-                    syncStore.markPublished(.rideHistory, at: event.createdAt)
-                }
-            } catch {
-                // Will retry on next reconnect
-            }
-        }
-    }
+    // MARK: - Identity Replacement
 
     private func prepareForIdentityReplacement(clearPersistedSyncState: Bool) async {
+        // 1. Connection
         connectionCoordinator.stop()
+
+        // 2. Ride + relay
         await rideCoordinator?.stopAll()
         await relayManager?.disconnect()
 
-        let syncStore = roadflareSyncStore
-        roadflareSyncStore = nil
-        if clearPersistedSyncState {
-            syncStore?.clearAll()
-        }
+        // 3. Sync (detaches ALL callbacks before any clearAll)
+        syncCoordinator?.teardown(clearPersistedState: clearPersistedSyncState)
+        syncCoordinator = nil
 
-        detachRepositoryTrackingCallbacks()
+        // 4. Repository data (callbacks already nil'd by teardown)
         driversRepository?.clearAll()
         if driversRepository == nil {
             driversPersistence.saveDrivers([])
@@ -691,11 +374,13 @@ final class AppState {
         rideHistory.clearAll()
         savedLocations.clearAll()
         settings.clearAll()
-        profileBackupSettingsTemplate = SettingsBackupContent()
         RideStatePersistence.clear()
+
+        // 5. UI state
         requestRideDriverPubkey = nil
         selectedTab = 0
 
+        // 6. Nil service refs
         rideCoordinator = nil
         relayManager = nil
         roadflareDomainService = nil
@@ -704,5 +389,4 @@ final class AppState {
         remoteConfigManager = nil
         bitcoinPrice.stop()
     }
-
 }
